// src/core/http/generated/models/tenantBlock.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { Tenant } from './tenant.ts';

export interface TenantBlock {
  id: string;
  tenant_id: string;
  notification_template_id: string;
  /** @nullable */
  block_reason?: string | null;
  created_at?: string;
  updated_at?: string;
  tenant?: Tenant;
}