// src/core/http/generated/models/updateNotificationTemplatePayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateNotificationTemplatePayload {
  key?: string;
  name?: string;
  description?: string;
  module?: string;
  event?: string;
  channel?: string;
  send_type?: string;
  subject?: string;
  body?: string;
  variables?: string[];
  is_active?: boolean;
  is_default?: boolean;
}